refactor: Hardened Organization Search, download remote files

+++ Packages/OrganizationSearch/Sources/OrganizationSearch/Structures/SearchResults.swift
@@ -10,16 +10,16 @@ import Foundation
 public struct SearchResults: Decodable, Hashable, Sendable {
 	
 	/// The total number of matching organizations in the index.
-	public let count: Double
+	public let count: Int
 
 	/// The list of individual search results for the current page.
 	public let hits: [SearchResult]
-	
+
 	/// Creates a new `SearchResults`.
 	/// - Parameters:
 	///   - count: The total number of matching organizations.
 	///   - hits: The individual search results for the current page.
-	public init(count: Double, hits: [SearchResult]) {
+	public init(count: Int, hits: [SearchResult]) {
 		self.count = count
 		self.hits = hits
 	}
